src/Eval-Runner/SXG.EvalPlatform.EvaluationEngine/requirements.txt
# Production dependencies for SXG Evaluation Platform
# Azure SDK packages - core dependencies
azure-storage-queue
azure-storage-blob  
azure-identity

# Azure AI Evaluation package
azure-ai-evaluation

# HTTP client for API calls
aiohttp

# Logging utilities
structlog
python-json-logger

# OpenTelemetry telemetry (modern standard)
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-logging
azure-monitor-opentelemetry-exporter